Deploy Swift to be used as Glance backend
In case of multi controller setup, Glance service is running on multiple
nodes. Hence, we can't use file store for images as it will be stored on
one host ans is absent on another, leading to VMs boot failures time to
time. So let's deploy Swift and use it as backend for Glance and solve
this issue.

Please note, an external Volume must be attached manually.